Implemented missing data workaround in breeding report
