+++ src/lib/reporting.ts
@@ -1,6 +1,21 @@
-import { BreedingReportData } from './actions.js';
+import { BreedingReportData, BreedingReportMode } from './actions.js';
 import { api } from './harnessnation.js'
-import { parseInt, sleep, toPercentage } from './utils.js';
+import { getRaces } from './races.js';
+import { formatEarnings, parseInt, toPercentage } from './utils.js';
+
+export interface BreedingReport {
+    totalFoals: number;
+    totalStarters: number;
+    totalWinners: number;
+    totalEarnings: number;
+    stakeStarters: number;
+    stakeWinners: number;
+    stakeStarts: number;
+    stakeWins: number;
+    stakePlaces: number;
+    stakeShows: number;
+    stakeEarnings: number;
+}
 
 type BreedingReportRow = (string | number | undefined)[];
 
@@ -9,7 +24,7 @@ type BreedingReportRow = (string | number | undefined)[];
  * @param {BreedingReportData} data - an object containing the ids of the horses and any header override.
  * @returns {Promise<string>} A `Promise` that resolves with the report as a base64-encoded data URI.
  */
-export async function generateBreedingReport({ ids, headers }: BreedingReportData): Promise<string> {
+export async function generateBreedingReport({ ids, headers, mode }: BreedingReportData): Promise<string> {
     const report: BreedingReportRow[] = [
         [
             'ID',
@@ -46,7 +61,7 @@ export async function generateBreedingReport({ ids, headers }: BreedingReportDat
     let batch: number[] | null;
 
     while ((batch = _ids.splice(0, 10)) && batch.length > 0) {
-        const rows = await Promise.all(batch.map(id => getBreedingReportRow(id)));
+        const rows = await Promise.all(batch.map(id => getBreedingReportRow(id, mode)));
         report.push(...rows.map(row => row.slice(0, report[0].length)));
     }
 
@@ -58,23 +73,90 @@ export async function generateBreedingReport({ ids, headers }: BreedingReportDat
             .join('\n'))}`;
 }
 
-
 /**
- * Generates a breeding report row for the given horse.
+ * Gets the breeding report for the given horse.
  * @param {number} id - the id of the horse.
- * @returns {Promise<BreedingReportRow>} A `Promise` that resolves with an array of row data values.
+ * @param {BreedingReportMode} mode - the lookup mode use for generating the report.
+ * @returns {Promise<BreedingReport>} A `Promise` that resolves with the breeding report data.
  */
-async function getBreedingReportRow(id: number): Promise<BreedingReportRow> {
+async function getBreedingReport(id: number, mode?: BreedingReportMode): Promise<BreedingReport> {
     const [profile, report] = await Promise.all([
         api.getHorse(id),
         api.getProgenyReport(id),
     ]);
 
-    const totalFoals = parseInt(profile.match(/<b[^>]*>\s*Total\s+Foals\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0');
-    const totalStarters = parseInt(report.match(/<b[^>]*>\s*Total\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0');
-    const totalWinners = parseInt(report.match(/<b[^>]*>\s*Total\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0');
-    const stakeStarters = parseInt(report.match(/<b[^>]*>\s*Stake\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0');
-    const stakeWinners = parseInt(report.match(/<b[^>]*>\s*Stake\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0');
+    const [stakeStarts, stakeWins, stakePlaces, stakeShows, stakeEarnings] = (
+        report.match(/<b[^>]*>\s*Stake\s+Results\s*:\s*<\/b[^>]*>\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*\(\$([\d,]+(?:\.\d+)?)\)/is)?.slice(1)
+        ?? ['0', '0', '0', '0', '0']
+    ).map(val => parseInt(val.replace(/,/g, '')));
+
+    const data: BreedingReport = {
+        totalFoals: parseInt(profile.match(/<b[^>]*>\s*Total\s+Foals\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalStarters: parseInt(report.match(/<b[^>]*>\s*Total\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalWinners: parseInt(report.match(/<b[^>]*>\s*Total\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalEarnings: parseInt(report.match(/<b[^>]*>\s*Total\s+Earnings\s*:\s*<\/b[^>]*>\s*\$([\d,]+(?:\.\d+)?)/is)?.[1] ?? '0'),
+        stakeStarters: parseInt(report.match(/<b[^>]*>\s*Stake\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        stakeWinners: parseInt(report.match(/<b[^>]*>\s*Stake\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        stakeStarts,
+        stakeWins,
+        stakePlaces,
+        stakeShows,
+        stakeEarnings,
+    };
+
+    if (mode === 'enhanced') {
+        const csrfToken = await api.getCSRFToken();
+        const progeny = await api.getProgenyList(id, csrfToken);
+
+        const candidates = Array.from(progeny
+            .replace(/\s*data-[^\s=]+="[^"]*"/g, '')
+            .matchAll(/<tr[^>]*>[\s\S]*?<\/tr[^>]*>/g)
+        )
+            .filter(([match]) => /<td[^>]*>\s*(1\d|[2-9])\s*<\/td[^>]*>\s*<td[^>]*>[\s\S]*?<\/td[^>]*>\s*<td[^>]*>[\s\S]*?<\/td[^>]*>\s*<td[^>]*>[\s\S]*?<\/td[^>]*>\s*<td[^>]*>\s*0\s*-\s*0\s*-\s*0\s*-\s*0\s*<\/td[^>]*>\s*<td[^>]*>\s*\$0\s*<\/td[^>]*>/i.test(match))
+            .map(([match]) => parseInt(/<a(?=[^>]*\bhorseLink\b)[^>]*\/horse\/(\d+)[^>]*>/i.exec(match)?.[1] ?? ''))
+            .filter(id => !Number.isNaN(id));
+
+        const chunkSize = 3;
+
+        for (let i = 0; i < candidates.length; i += chunkSize) {
+            const chunk = candidates.slice(i, i + chunkSize);
+
+            for (const races of await Promise.all(chunk.map(id => getRaces(id)))) {
+                if (races.length < 1)
+                    continue;
+
+                data.totalStarters++;
+                data.totalEarnings += races.getEarnings();
+
+                if (races.some(r => r.finish === 1))
+                    data.totalWinners++;
+
+                const [starts, wins, places, shows] = races.getSummary(r => r.stake === true && r.elim === false);
+                data.stakeStarters += (starts > 0 ? 1 : 0);
+                data.stakeWinners += (wins > 0 ? 1 : 0);
+                data.stakeStarts += starts;
+                data.stakeWins += wins;
+                data.stakePlaces += places;
+                data.stakeShows += shows;
+                data.stakeEarnings += races.getEarnings(r => r.stake === true && r.elim === false);
+            }
+        }
+    }
+
+    return data;
+}
+
+/**
+ * Generates a breeding report row for the given horse.
+ * @param {number} id - the id of the horse.
+ * @param {BreedingReportMode} mode - the data lookup mode.
+ * @returns {Promise<BreedingReportRow>} A `Promise` that resolves with an array of row data values.
+ */
+async function getBreedingReportRow(id: number, mode?: BreedingReportMode): Promise<BreedingReportRow> {
+    const [profile, report] = await Promise.all([
+        api.getHorse(id),
+        getBreedingReport(id, mode),
+    ]);
 
     return [
         +id,
@@ -88,17 +170,127 @@ async function getBreedingReportRow(id: number): Promise<BreedingReportRow> {
         profile.match(/<b[^>]*>\s*Dam-Sire\s*:\s*<\/b[^>]*>\s*<a[^>]*>(.*?)<\/a[^>]*>/is)?.[1]?.trim() ?? 'Unknown',
         profile.match(/<b[^>]*>\s*Lifetime\s+Race\s+Record\s*<\/b[^>]*>\s*<br[^>]*>\s*(.*?)<br[^>]*>/is)?.[1]?.trim(),
         profile.match(/<b[^>]*>\s*Stake\s+Record\s*<\/b[^>]*>\s*<br[^>]*>\s*(.*?)<br[^>]*>/is)?.[1]?.trim(),
-        totalFoals,
-        totalStarters,
-        toPercentage(totalStarters, totalFoals),
-        totalWinners,
-        toPercentage(totalWinners, totalStarters),
-        report.match(/<b[^>]*>\s*Total\s+Earnings\s*:\s*<\/b[^>]*>\s*([$\d,]+(?:\.\d+)?)/is)?.[1] ?? '$0',
-        report.match(/<b[^>]*>\s*Average\s+Earnings\s+per\s+Starter\s*:\s*<\/b[^>]*>\s*([$\d,]+(?:\.\d+)?)/is)?.[1] ?? '$0',
-        stakeStarters,
-        toPercentage(stakeStarters, totalStarters),
-        stakeWinners,
-        toPercentage(stakeWinners, totalStarters),
-        ...(report.match(/<b[^>]*>\s*Stake\s+Results\s*:\s*<\/b[^>]*>\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*\(([$\d,]+(?:\.\d+)?)\)/is)?.slice(1) ?? ['0', '0', '0', '0', '$0']),
+        report.totalFoals,
+        report.totalStarters,
+        toPercentage(report.totalStarters, report.totalFoals),
+        report.totalWinners,
+        toPercentage(report.totalWinners, report.totalStarters),
+        formatEarnings(report.totalEarnings),
+        formatEarnings(report.totalStarters > 0 ? report.totalEarnings / report.totalStarters : 0),
+        report.stakeStarters,
+        toPercentage(report.stakeStarters, report.totalStarters),
+        report.stakeWinners,
+        toPercentage(report.stakeWinners, report.totalStarters),
+        report.stakeStarts,
+        report.stakeWins,
+        report.stakePlaces,
+        report.stakeShows,
+        formatEarnings(report.stakeEarnings),
     ];
+}
+
+async function getProgenyIds(id: number): Promise<Set<number>> {
+    return new Set(await api.getProgenyList(id).then(html =>
+        html.matchAll(/<td[^>]*>\s*<a[^>]*\/horse\/(\d+)[^>]*><span[^>]*>(.*?)<\/span[^>]*><\/a[^>]*>/).map(([match, id]) => parseInt(id))
+    ));
+}
+
+async function summarizeBrokenProgeny(ids: Array<number>): Promise<{
+    totalStarters: number;
+    totalWinners: number;
+    totalEarnings: number;
+    stakeStarters: number;
+    stakeWinners: number;
+    stakeStarts: number;
+    stakeWins: number;
+    stakePlaces: number;
+    stakeShows: number;
+    stakeEarnings: number;
+}> {
+    const token = ids.length > 0 ? await api.getCSRFToken() : '';
+
+    let totalStarters: number = 0,
+        totalWinners: number = 0,
+        totalEarnings: number = 0,
+        stakeStarters: number = 0,
+        stakeWinners: number = 0,
+        stakeStarts: number = 0,
+        stakeWins: number = 0,
+        stakePlaces: number = 0,
+        stakeShows: number = 0,
+        stakeEarnings: number = 0;
+
+    for (const races of await Promise.all(ids.map(id => getRaces(parseInt(id), token)))) {
+        if (races.length < 1)
+            continue;
+
+        totalStarters++;
+        totalEarnings += races.getEarnings();
+
+        if (races.some(r => r.finish === 1))
+            totalWinners++;
+
+        const [starts, wins, places, shows] = races.getSummary(r => r.stake === true && r.elim === false);
+        stakeStarters += (starts > 0 ? 1 : 0);
+        stakeWinners += (wins > 0 ? 1 : 0);
+        stakeStarts += starts;
+        stakeWins += wins;
+        stakePlaces += places;
+        stakeShows += shows;
+        stakeEarnings += races.getEarnings(r => r.stake === true && r.elim === false);
+    }
+
+    return {
+        totalStarters,
+        totalWinners,
+        totalEarnings,
+        stakeStarters,
+        stakeWinners,
+        stakeStarts,
+        stakeWins,
+        stakePlaces,
+        stakeShows,
+        stakeEarnings,
+    };
+}
+
+async function summarizeProgenyReport(id: number): Promise<{
+    totalFoals: number;
+    totalStarters: number;
+    totalWinners: number;
+    totalEarnings: number;
+    stakeStarters: number;
+    stakeWinners: number;
+    stakeStarts: number;
+    stakeWins: number;
+    stakePlaces: number;
+    stakeShows: number;
+    stakeEarnings: number;
+}> {
+    const report = await api.getProgenyReport(id);
+
+    const [
+        stakeStarts = 0,
+        stakeWins = 0,
+        stakePlaces = 0,
+        stakeShows = 0,
+        stakeEarnings = 0,
+    ] = (
+        report.match(/<b[^>]*>\s*Stake\s+Results\s*:\s*<\/b[^>]*>\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*-\s*([\d,]+)\s*\(\$([\d,]+(?:\.\d+)?)\)/is)?.slice(1)
+        ?? []
+    ).map(val => parseInt(val));
+
+    return {
+        totalFoals: parseInt(report.match(/<b[^>]*>\s*Total\s+Foals\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalStarters: parseInt(report.match(/<b[^>]*>\s*Total\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalWinners: parseInt(report.match(/<b[^>]*>\s*Total\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        totalEarnings: parseInt(report.match(/<b[^>]*>\s*Total\s+Earnings\s*:\s*<\/b[^>]*>\s*\$([\d,]+(?:\.\d+)?)/is)?.[1] ?? '0'),
+        stakeStarters: parseInt(report.match(/<b[^>]*>\s*Stake\s+Starters\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        stakeWinners: parseInt(report.match(/<b[^>]*>\s*Stake\s+Winners\s*:\s*<\/b[^>]*>\s*([\d,]+)/is)?.[1] ?? '0'),
+        stakeStarts,
+        stakeWins,
+        stakePlaces,
+        stakeShows,
+        stakeEarnings,
+    }
 }
